+++ R/cf_query.R
@@ -5,6 +5,7 @@
 #' @param body Additional parameters to add to request body.
 #' @param key A character string containing a Crowdflower API key.
 #' @param type A character string specifying the type of HTTP request (\dQuote{GET}, \dQuote{POST}, \dQuote{PUT}, etc.).
+#' @param parse A logical indicating whether to parse the response (\code{TRUE}, the default) or return a raw vector (\code{FALSE}).
 #' @param base_url A character string specifying the base URL of API request.
 #' @param ... Additional arguments passed to an HTTP request function (e.g, \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{PUT}}).
 #' @return A character string containing the content of the HTTP request.
@@ -14,6 +15,7 @@ cf_query <- function(endpoint,
                     query = NULL, 
                     body = NULL, 
                     type = "GET", 
+                    parse = TRUE,
                     key = Sys.getenv("CROWDFLOWER_API_KEY"),
                     base_url = 'https://api.crowdflower.com/v1/', 
                     ...){
@@ -33,20 +35,23 @@ cf_query <- function(endpoint,
     httr::stop_for_status(r)
     
     # trying to parse response
-    error <- tryCatch(httr::content(r),
-                      error = function(e) e)
-    if (!inherits(error, 'error')){
-        # if JSON contains embedded error message, return it
-        if (length(error$error)){
-            stop(error$error)
+    if (isTRUE(parse)) {
+        error <- tryCatch(httr::content(r),
+                          error = function(e) e)
+        if (!inherits(error, 'error')) {
+            # if JSON contains embedded error message, return it
+            if (length(error$error)){
+                stop(error$error)
+            }
+            # else return the response value
+            response <- error
+        } else {
+            # try returning raw text response
+            warning("Response was not JSON")
+            response <- httr::content(r, 'text')
         }
-        # else return the response value
-        response <- error
     } else {
-        # try returning raw text response
-        warning("Response was not JSON")
-        response <- httr::content(r, 'text')
+        response <- httr::content(r, "raw")
     }
-
     return(response)
 }